pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "star-hash"
version = "0.1.0"
description = "A tool to generate celestial stamps encoding unique spacetime coordinates."
authors = [
  { name="Gabriele Pizzi", email="info@gabrielepizzi.com" },
]
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "astronomy-engine",
    "typer[all]",
    "svgwrite",
    "qrcode",
    "qrcode",
]

[project.scripts]
star-hash = "star_hash.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["star_hash*"]
